requirements.txt
celery>=5.3.6
flower>=2.0.1
alembic>=1.12.1
asyncpg>=0.29.0
SQLAlchemy>=2.0.23
python-dotenv>=1.0.0
fastapi[all]>=0.104.1
fastapi-cache2[redis]>=0.2.1
fastapi-users[sqlalchemy]>=6.0.1